Replace :ref: with :doc: wherever possible
ReStructuredText provides many ways to create internal links
(i.e., cross-referencing specific locations throughout a project).
The two most common ones are :doc: (for linking to a specific page) and
:ref: (for linking to any arbitrary location).

:ref: requires you to create a destination marker
somewhere in the project, like so:

    .. foo::

so that it can be linked thusly:

    :ref:`foo`

:doc:, on the other hand, simply takes the path of a page in the
project, and requires no additional marker to be added anywhere:

    :doc:`/path/to/page`

Prior to this commit, there were many :ref: links that simply pointed to
the top of a target page. :doc: works just as well for this purpose,
so this commit replaces all those instances of :ref: with :doc:.