Create "Misc" TOC section for Composer Settings docs
Originally, we had a docs page for "Settings in Tickets"
which referred the ticket composer settings dialog.

This page had no good place in the TOC
(it was originally placed under the "Channels > Email" page),
and the name was confusing
(since we already have a "Settings > Ticket" page).

This commit renames it to "Composer Settings"
and places it under a new TOC section, "Misc".