Fix RST indentation
Python convention (by which I mean the style guide
for official Python language documentation)
is to use three spaces for indentation in ReStructuredText files.

This is not a strict requirement of RST syntax[0],
but it produces a clean visual structure for RST's directive syntax:

    .. note:: Lorem ipsum dolor

       sit amet consectetur

In the case of lists, however, two spaces look better:

    - Lorem ipsum dolor
      sit amet consectetur

but in all other cases, we use three spaces for maximum consistency.

This commit replaces all indentation according to the rules above.

[0]: https://stackoverflow.com/questions/48311159