Fix German-style hyphens and capitalization
The existing docs contained many instances of German-style hyphens and
capitalization. In English, capitalization is reserved for titles and
proper nouns.

Of course, in technical writing, sometimes it's appropriate to
capitalize words that refer to elements on the screen, such as:

> When you're finished writing your message, click "Submit".

In these cases, the screen element ("Submit") should be emphasized with
additional styling, such as quotation marks, italics, or boldfaced
lettering.

(I don't have a source for this, unfortunately, but I would argue that
it's more than my personal opinion.)

This commit doesn't add any such styling to these cases, but it does
fix capitalization where appropriate.

Hyphens are used sparingly, and the rules are confusing. "Auto-reply" is
okay, but "API token" is not (because "API" is its own word, but "auto"
is not). "Email-related" is also okay, because it forms one single
adjective.