Update trigger documentary (#48)
* initial rework

This rework -for now- mainly contains structural changes and preparations.

* adding toctree and links

* finished wording on initial trigger page

* change image to figure for better spacings on example page

* Update trigger pictures

This replaces old jpgs with pngs (as suggested by Ryan) and now uses a subfolder for better overview.

* add hint for triggers within calendars

As triggers now also can rely on calendars, I've added a small hint.

* rework images on conditions and actions for triggers

This meets the suggested pngs of ryan and also shows the current UI look and feel.

* slight rework and addition

Slightly reworked existing texts and added further things like execution time condition for triggers.

* Enhance calendar wording for it's cope

Accepting Ryans suggestion on enhance the messy wording on the calendar introduction. This will change it into a way better listing and clear the goal up a lot.

Co-Authored-By: Ryan Lue <hello@ryanlue.com>

* re arrange navigation structure and places where to find information

As per internal discussion with ryan, as we were pushing rst capatilities a bit too far and thus ran into issues with navigation urls not fitting the toctree as we wanted to.

For this we moved information to sub pages.

* add introduction text

The learn by example page jumped in directly. It now comes with an introduction text to help the reader better.

* add postmaster filter als other alternative to triggers

In limitations we're talking about possible issues you can't solve with triggers. Schedulers usually are the correct alternative here, however, in some situations a postmaster filter is a good choice as well. This is why we shouldn't leave it out.

* Rename "other notifications" to "system notifications"

As this was basically always about system notifications, this is an alot better wording.

* Added limitations hint for the condition and actions page

This helps to understand the user the scope and limitations of triggers better.

* Flesh out and restructure Trigger intro page

* Highlight important information with hint/note insets

* Rename "Conditions and Actions" page for clarity

* Rename Trigger subsection files to match their titles

Co-authored-by: Ryan Lue <hello@ryanlue.com>