Introducing S/MIME integration (#57)
* WIP: SMIME documentation

* WIP: SMIME documentation

* WIP: SMIME documentation

This commit adds screenshots from Zammad, introduces a small "common issues" section and provides description on what to do to enable SMIME.

* adding further possible error case

This should help administrators experiencing the same issue.

* adding hint for later added certificates

If S/MIME integration has been activated and a mail is being received Zammad can either not decrypt or verify the signed message, it will display a warning. This warning allows the user to re-verify the content on a later moment.

* adding not for further scope

Adding a note for admins that those issues may also affect triggers or schedulers.

* Match image filepaths to their articles

This commit was undertaken in preparation for PR #57.[0]

PR #57 adds a new article (system/integrations/smime.rst),
which naturally includes new images.
The filepaths of images included in PR #57
will be logically better-organized than those for existing articles.

In the interest of consistency,
those images have been re-organized here
to match what will be introduced next.

[0]: https://github.com/zammad/zammad-admin-documentation/pull/57

* Adding information to logging functionality for SMIME

* Update trigger & scheduler to link to their documentation pages

To help admins to have a quick jump to the right place.

* fix spacings

fix 2 spacings to 3 spacings.

* fixed typo (missing R)

* Add further "what might go terribly wrong" for S/MIME

During my own playing around I encountered a further possible issue which Id ocumented here.

* restructure text for certificates and replace image(s)

This commit orientates itself a bit on the email channel documentation which provides a screencast for a quick overview of what can be done in the described part of the page.

The text has been rearranged to make that step more uniform.

* Introduce a small info text to default behavior for smime

This functionality came a bit later than the docs where written. This also clears the configuration part further up a bit better.

* replace image for smime log with current implementation

* Match image filepaths to their articles

* Compress & re-crop image assets

* QA/Rewrite S/MIME integration article

* cherry pick zabbix fix

* Add further error message if trying to supply anything but PEM

Co-authored-by: Ryan Lue <hello@ryanlue.com>