Prep for refactoring of new Google channel docs
Since the Google channel is basically a specialized email channel,
there's a _lot_ of overlap between the Email and Google docs.

The original approach to managing this overlap was:

1. if a **fragment of a page** needed to look the same,
   duplicate it; and
2. if a **whole page** needed to look the same,
   link directly to it.

This approach created two problems:

1. not just unnecessary duplication,
   but the risk of duplicated sections falling out of sync
   when changes are made in the future; and
2. issues with "Next" / "Previous" button navigation
   when the Table of Contents for the Google section
   referenced a page from the Email section.

The next commit will address these issues
by leveraging the "include" directive.[0]
This commit prepares for this change by rearranging the file structure
so that "including" part of one file in another
produces the appropriate :doc: links.

[0]: https://docutils.sourceforge.io/docs/ref/rst/directives.html#include