Introduce google channel (#58)
* Introduce google channel

This commit introduces the google channel for Zammad.

It makes heavy use of already existing pages of the email channel, as they are technically identically (better: the *same*) and thus don't need double work in my opinion.

I orientated myself as much as I could on the email channel.

* adding important note on left over mail addresses

You can't add gmail accounts if one of the email addresses of that account is already known to Zammad (no matter as an account or leftover mail address)

* Explain why we instantly remove google mail aliases with account removal

* Prep for refactoring of new Google channel docs

Since the Google channel is basically a specialized email channel,
there's a _lot_ of overlap between the Email and Google docs.

The original approach to managing this overlap was:

1. if a **fragment of a page** needed to look the same,
   duplicate it; and
2. if a **whole page** needed to look the same,
   link directly to it.

This approach created two problems:

1. not just unnecessary duplication,
   but the risk of duplicated sections falling out of sync
   when changes are made in the future; and
2. issues with "Next" / "Previous" button navigation
   when the Table of Contents for the Google section
   referenced a page from the Email section.

The next commit will address these issues
by leveraging the "include" directive.[0]
This commit prepares for this change by rearranging the file structure
so that "including" part of one file in another
produces the appropriate :doc: links.

[0]: https://docutils.sourceforge.io/docs/ref/rst/directives.html#include

* Reduce duplication and fix navigation using ..include::

* Clarify difference between Google/Email pages

* Convert DOS line endings (CRLF) to unix (LF)

* Finalize Google > Accounts page

* Revise Google > Account Setup page

* Revise Google > Register OAuth App page

* Reduce duplication in Google  pages with include:: directive

* Bring back in deletion information for aliases

As per discussion with Ryan, this adds the information of aliases being removed back in.

* add notification to for system notifications

I thought that "use a regular email channel" with linking to "email accounts page" which presents the notification channel as last option might be confusing.

To ensure people not adding email channels and then wondering why "notifications don't work", I added "notification"  to "regular email channel" which should make that crystal clear.

* adjust mail server settings and fetch behavior

Google channel technically doesn't know mail server settings. The screenshot and behavior "on edit" is different which should now be clearer by this commit.

* new export of screencast for better quality

Earlier exports where quite bad for their size and were incorrectly showing some of the color information. This doesn't help if you're trying to recognize UIs.

This commit fixes this issue.
One images is missing on that part.

* Fix HTML compilation bug from include:: directive

* fix gif quality for adding a gmail account to Zammad

Co-authored-by: Ryan Lue <hello@ryanlue.com>