Maintenance: Fix 2020 emoji & add search function hint
Earlier version of this file was using a very current 2020 emoji which isn't available on a broad number of systems yet (magic wand). To improve compatibility we changed it to sparkle.

Also the hints on how an agent can adjust users was incomplete (search function was missing). In my opinion this is one of the most used ways and thus should be declared. This also helps admins to tell their users the options the agents have.