Maintenance: Re-add preference info & adjust ldap wording
A common confusion is that a customer gets the "notification" permission, but doesn't receive any system notification. To make this clearer from the beginning, I've added that one liner. (Flew below my radar earlier)

Regarding password permissions:
Changing the user password when using LDAP does only "visually" drift passwords. In fact it actually sets a second password (a local one) which allows you to use two passwords, one local and one from your ldap source.

As Zammad does not change the ldap password for you, this will be confusing. This is why I adjusted the end of the sentence.