Enhance information regarding notification channel (#74)
* Replace miss leading wording regarding notification sender address

Before the notification channel was referenced as being relevant for automatic replies (e.g. trigger based). This leads admins in the wrong direction.

The notification channel is for system notifications only. This commit fixes this.

* Add hint regarding sender address

The documentation misses the hint that the sender address for a notification channel is to be configured else where.

While this is confusing enough, the missing hint doesn't catch affected users.

This came up through https://community.zammad.org/t/untrapped-exception-in-pythonfilter/5323/8

* Update channels/email/accounts/email-notification.rst

Co-authored-by: Ryan Lue <hello@ryanlue.com>

Co-authored-by: Ryan Lue <hello@ryanlue.com>