Address PR71 refactor (small changes, additions) (#77)
* Maintenance: Adjust Screenshots to "new" look and feel

* Maintenance: Fix 2020 emoji & add search function hint

Earlier version of this file was using a very current 2020 emoji which isn't available on a broad number of systems yet (magic wand). To improve compatibility we changed it to sparkle.

Also the hints on how an agent can adjust users was incomplete (search function was missing). In my opinion this is one of the most used ways and thus should be declared. This also helps admins to tell their users the options the agents have.

* Maintenance: Be a bit more specific

In order to solve the question "why don't I get notifications" ?

Earlier this was mentioned in within "the standard issue" on the example part:

"This provides the complete notification and ticket experience for the group in question."

Above should be enough though.

* Maintenance: Replace links to functionalities

Linking to administration documentation for permissions that are about agents is confusing in my opinion. Thus we're rather linking to the user documentation and provide hints on "hey, you gotta configure stuff first here".

The only exception is the reporting currently, because this needs work and love first.

* Maintenance: Readd "users are global" information

This is a common question on "how to restrict" users being available globally.

As this is not a function within Zammad, we should state it explicitly. This also fixes #46

* Maintenance: Re-add preference info & adjust ldap wording

A common confusion is that a customer gets the "notification" permission, but doesn't receive any system notification. To make this clearer from the beginning, I've added that one liner. (Flew below my radar earlier)

Regarding password permissions:
Changing the user password when using LDAP does only "visually" drift passwords. In fact it actually sets a second password (a local one) which allows you to use two passwords, one local and one from your ldap source.

As Zammad does not change the ldap password for you, this will be confusing. This is why I adjusted the end of the sentence.

* Address Ryans suggestions

As per comments on PR#77

* Move user being global hint to groups as suggested by Ryan

* Update manage/roles/user-preferences-permissions.rst

Use Ryans suggestion to raise not too many questiosn but make clear that when using LDAP you should not allow affected users to set or change their password.

Co-authored-by: Ryan Lue <hello@ryanlue.com>

Co-authored-by: Ryan Lue <hello@ryanlue.com>