Replace auto-collapsing code block with y-scroll
Prior to this commit, this repo used JavaScript
to identify long code blocks (el.clientHeight > 250px),
automatically collapse them,
and add a child button to toggle their collapsed state.

This added considerable complexity to the page,
which can be roughly approximated by setting

    y-overflow: scroll;

on the code block instead.

=== Tradeoff

Previously, JS would automatically detect
which code blocks should be collapsed.
This was good in the sense that it was automatic,
but bad in the sense that it would, for instance,
collapse code blocks 251px tall
(meaning clicking on these code blocks would expand a single pixel).

Now, long code blocks must be manually identified
and tagged by wrapping them in a container directive:

    .. container:: long-code-block

       .. code-block::

          ...